Remove batbox::future_ext
