+++ crates/geng/src/asset/native.rs
@@ -14,13 +14,25 @@ impl AssetManager {
             threadpool: default(),
         }
     }
+    fn spawn<T: Send + 'static, F: FnOnce() -> T + Send + 'static>(
+        &self,
+        f: F,
+    ) -> futures::channel::oneshot::Receiver<T> {
+        let (sender, receiver) = futures::channel::oneshot::channel();
+        self.threadpool.execute(move || {
+            if sender.send(f()).is_err() {
+                panic!("Failed to send value");
+            }
+        });
+        receiver
+    }
 }
 
 impl LoadAsset for ugli::Texture {
     fn load(geng: &Geng, path: &std::path::Path) -> AssetFuture<Self> {
         let ugli = geng.ugli().clone();
         let path = path.to_owned();
-        let image_future = geng.inner.asset_manager.threadpool.spawn(move || {
+        let image_future = geng.inner.asset_manager.spawn(move || {
             debug!("Loading {:?}", path);
             fn load(path: &std::path::Path) -> anyhow::Result<image::RgbaImage> {
                 let image = image::open(path).context(format!("Failed to load {:?}", path))?;
@@ -41,16 +53,15 @@ impl LoadAsset for Sound {
     fn load(geng: &Geng, path: &std::path::Path) -> AssetFuture<Self> {
         let geng = geng.clone();
         let path = path.to_owned();
-        let data =
-            geng.inner
-                .asset_manager
-                .threadpool
-                .spawn(move || -> Result<_, anyhow::Error> {
-                    debug!("Loading {:?}", path);
-                    let mut data = Vec::new();
-                    std::fs::File::open(path)?.read_to_end(&mut data)?;
-                    Ok(data)
-                });
+        let data = geng
+            .inner
+            .asset_manager
+            .spawn(move || -> Result<_, anyhow::Error> {
+                debug!("Loading {:?}", path);
+                let mut data = Vec::new();
+                std::fs::File::open(path)?.read_to_end(&mut data)?;
+                Ok(data)
+            });
         Box::pin(async move { Ok(Sound::new(&geng, data.await??)) })
     }
     const DEFAULT_EXT: Option<&'static str> = Some("wav");
